Move warehouse dtos to new format
