+++ apps/api/src/models/organizations/dto/create-organization.dto.ts
@@ -1,6 +1,7 @@
 import { Type } from 'class-transformer';
 import { Length, ValidateNested } from 'class-validator';
-import { CreateWarehouseDto, ICreateOrganizationDto } from 'shared-types';
+import { ICreateOrganizationDto } from 'shared-types';
+import { CreateWarehouseDto } from '../../warehouses/dto/create-warehouse.dto';
 
 export class CreateOrganizationDto implements ICreateOrganizationDto {
 	@Length(2, 32)
