+++ apps/api/src/models/warehouses/dto/create-warehouse.dto.ts
@@ -1,0 +1,11 @@
+import { IsOptional, Length } from 'class-validator';
+import { ICreateWarehouseDto } from 'shared-types';
+
+export class CreateWarehouseDto implements ICreateWarehouseDto {
+	@Length(2, 32)
+	name: string;
+
+	@IsOptional()
+	@Length(2, 32)
+	address?: string;
+}
