+++ apps/api/src/models/warehouses/dto/update-warehouse.dto.ts
@@ -1,0 +1,4 @@
+import { CreateWarehouseDto } from './create-warehouse.dto';
+import { IUpdateWarehouseDto } from 'shared-types';
+
+export class UpdateWarehouseDto extends CreateWarehouseDto implements IUpdateWarehouseDto {}
