+++ apps/api/src/models/warehouses/warehouses.controller.ts
@@ -12,7 +12,7 @@ import {
 } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 import { Types } from 'mongoose';
-import { CreateWarehouseInOrgDto, PageDto, UpdateWarehouseDto, WarehouseDto } from 'shared-types';
+import { PageDto, WarehouseDto } from 'shared-types';
 import { ParseObjectIdPipe } from '../../pipes/prase-object-id.pipe';
 import { OrganizationsStatsService } from '../organizations/organizations-stats.service';
 import { OrganizationsService } from '../organizations/organizations.service';
@@ -20,6 +20,8 @@ import { Warehouse } from './schemas/warehouse.schema';
 import { WarehousesService } from './warehouses.service';
 import { PageQueryValidationPipe } from '../../pipes/page-query-validation.pipe';
 import { PageQueryDto } from '../../dto/page-query.dto';
+import { CreateWarehouseInOrgDto } from './dto/create-warehouse-in-org.dto';
+import { UpdateWarehouseDto } from './dto/update-warehouse.dto';
 
 @ApiTags('warehouses')
 @Controller('warehouses')
