+++ packages/shared-types/package.json
@@ -11,7 +11,5 @@
         "typescript": "^4.8.4"
     },
     "dependencies": {
-        "class-transformer": "^0.5.1",
-        "class-validator": "^0.14.0"
     }
 }
