+++ packages/shared-types/src/inventory/ICreateInventoryItemDto.ts
