--- packages/shared-types/src/warehouse/CreateWarehouseDto.ts
@@ -1,10 +1,0 @@
-import { IsOptional, Length } from "class-validator"
-
-export class CreateWarehouseDto {
-    @Length(2, 32)
-    name: string
-
-    @IsOptional()
-    @Length(2, 32)
-    address?: string
-}
