--- packages/shared-types/src/warehouse/CreateWarehouseInOrgDto.ts
@@ -1,12 +1,0 @@
-import { Type } from "class-transformer";
-import { IsMongoId, ValidateNested } from "class-validator";
-import { CreateWarehouseDto } from "./CreateWarehouseDto";
-
-export class CreateWarehouseInOrgDto {
-    @IsMongoId()
-    organizationId: string;
-
-    @ValidateNested()
-    @Type(() => CreateWarehouseDto)
-    warehouse: CreateWarehouseDto;
-}
