+++ packages/shared-types/src/warehouse/ICreateWarehouseDto.ts
@@ -1,0 +1,4 @@
+export interface ICreateWarehouseDto {
+    name: string
+    address?: string
+}
