+++ packages/shared-types/src/warehouse/ICreateWarehouseInOrgDto.ts
@@ -1,0 +1,6 @@
+import { ICreateWarehouseDto } from "./ICreateWarehouseDto";
+
+export class ICreateWarehouseInOrgDto {
+    organizationId: string;
+    warehouse: ICreateWarehouseDto;
+}
