+++ packages/shared-types/src/warehouse/IUpdateWarehouseDto.ts
@@ -1,0 +1,3 @@
+import { ICreateWarehouseDto } from "./ICreateWarehouseDto";
+
+export interface IUpdateWarehouseDto extends ICreateWarehouseDto { }
