--- packages/shared-types/src/warehouse/UpdateWarehouseDto.ts
@@ -1,3 +1,0 @@
-import { CreateWarehouseDto } from "./CreateWarehouseDto";
-
-export class UpdateWarehouseDto extends CreateWarehouseDto { }
